// src/scripts/interfaces/post.ts
// Interface automatically generated by schemas-to-ts


import type {Media} from "@/scripts/interfaces/defaults/Media";
import type {User, User_Plain} from "@/scripts/interfaces/defaults/User";
import type {
  AdminPanelRelationPropertyModification
} from "@/scripts/interfaces/defaults/AdminPanelRelationPropertyModification";

export interface Post {
  id: number;
  attributes: {
    createdAt: Date;    updatedAt: Date;    publishedAt?: Date;    postTitle: string;
    postSlug: string;
    postCover: { data: Media };
    postContent: string;
    postAuthor?: { data: User };
    locale: string;
    localizations?: { data: Post[] };
  };
}
export interface Post_Plain {
  id: number;
  createdAt: Date;  updatedAt: Date;  publishedAt?: Date;  postTitle: string;
  postSlug: string;
  postCover: Media;
  postContent: string;
  postAuthor?: User_Plain;
  locale: string;
  localizations?: Post[];
}

export interface Post_NoRelations {
  id: number;
  createdAt: Date;  updatedAt: Date;  publishedAt?: Date;  postTitle: string;
  postSlug: string;
  postCover: number;
  postContent: string;
  postAuthor?: number;
  locale: string;
  localizations?: Post[];
}

export interface Post_AdminPanelLifeCycle {
  id: number;
  createdAt: Date;  updatedAt: Date;  publishedAt?: Date;  postTitle: string;
  postSlug: string;
  postCover: AdminPanelRelationPropertyModification<Media>;
  postContent: string;
  postAuthor?: AdminPanelRelationPropertyModification<User_Plain>;
  locale: string;
  localizations?: Post[];
}