// src/scripts/interfaces/project.ts
// Interface automatically generated by schemas-to-ts


import type {Media} from "@/scripts/interfaces/defaults/Media";
import type {User, User_Plain} from "@/scripts/interfaces/defaults/User";
import type {
  AdminPanelRelationPropertyModification
} from "@/scripts/interfaces/defaults/AdminPanelRelationPropertyModification";

export enum ProjectStatus {
  Planned = 'Planned',
  Idle = 'Idle',
  Wip = 'WIP',
  Released = 'Released',
  Lts = 'LTS',
  Abandoned = 'Abandoned',}
export enum ProjectType {
  Game = 'Game',
  Software = 'Software',
  Website = 'Website',
  Other = 'Other',}

export interface Project {
  id: number;
  attributes: {
    createdAt: Date;    updatedAt: Date;    publishedAt?: Date;    projectName: string;
    projectDescription: string;
    projectSlug: string;
    projectCover: { data: Media };
    isInEarlyAccess: boolean;
    projectStatus: ProjectStatus;
    projectType: ProjectType;
    projectAuthor?: { data: User };
    locale: string;
    localizations?: { data: Project[] };
  };
}
export interface Project_Plain {
  id: number;
  createdAt: Date;  updatedAt: Date;  publishedAt?: Date;  projectName: string;
  projectDescription: string;
  projectSlug: string;
  projectCover: Media;
  isInEarlyAccess: boolean;
  projectStatus: ProjectStatus;
  projectType: ProjectType;
  projectAuthor?: User_Plain;
  locale: string;
  localizations?: Project[];
}

export interface Project_NoRelations {
  id: number;
  createdAt: Date;  updatedAt: Date;  publishedAt?: Date;  projectName: string;
  projectDescription: string;
  projectSlug: string;
  projectCover: number;
  isInEarlyAccess: boolean;
  projectStatus: ProjectStatus;
  projectType: ProjectType;
  projectAuthor?: number;
  locale: string;
  localizations?: Project[];
}

export interface Project_AdminPanelLifeCycle {
  id: number;
  createdAt: Date;  updatedAt: Date;  publishedAt?: Date;  projectName: string;
  projectDescription: string;
  projectSlug: string;
  projectCover: AdminPanelRelationPropertyModification<Media>;
  isInEarlyAccess: boolean;
  projectStatus: ProjectStatus;
  projectType: ProjectType;
  projectAuthor?: AdminPanelRelationPropertyModification<User_Plain>;
  locale: string;
  localizations?: Project[];
}